refactor: react & web-components folder structure
